Add files via upload
OK so as I sent you the pics trough text. I uploaded the " Free Template" and the Template from class that we were given as an option to start from. I want my page to resemble the other template. From the pictures I sent you, I want my header to float above the body or box or whatever it is, with the logo and title and h2 under the logo and then have the body or main content, w.e under the header. No border or anything around the header, like how it is in my horrible web page I started. Ive been trying to remove the border and have it float and resemble the other page but it hasn't been working. Hopefully I explained it right but you get what I mean and how I want it from the pictures I sent you through text messages.